pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tutorial" 
version = "0.1.0"
description = "Introductory fMRI tutorial package"
authors = [{name = "Yuanfang Zhao", email = "yzhao@jhu.edu"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "nilearn==0.12.1",
  "matplotlib>=3.8,<4.0",
  "pybids==0.20.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src/tutorial"]
exclude = ["data/**", "notebooks/**"]

[tool.hatch.build.targets.sdist]
exclude = ["data/**", "notebooks/**", ".venv/**", ".ipynb_checkpoints/**"]
